Fix TypeScript error in ethereumUtils.ts
Resolved the TypeScript error related to the argument type in the `sendEthereumTransaction` function. Ensured that the argument passed is compatible with the expected `BigNumberish` type.
[skip gpt_engineer]

+++ src/utils/networks/ethereumUtils.ts
@@ -32,7 +32,7 @@ export const getEthereumBalance = async (address: string): Promise<string> => {
       setTimeout(() => reject(new Error('Timeout')), 10000)
     );
     
-    const rawBalance = await Promise.race([balancePromise, timeoutPromise]);
+    const rawBalance = await Promise.race([balancePromise, timeoutPromise]) as bigint;
     const balance = ethers.formatEther(rawBalance);
     console.log('ETH balance:', balance);
     return balance;
